fix image

+++ app/Services/BotFilterMessageServices.php
@@ -23,42 +23,44 @@ class BotFilterMessageServices
    {
       $columnTable = $column == 'telegram_id' ? 'telegram' : 'vkontakte';
       if ($this->checkMute($user_id, $column)) {
-         if (!$sticker) {
-            $analyzeText = $this->analyzeText($text);
-         } else {
-            $analyzeText = [
-               'status' => true,
-               'type' => 'words'
-            ];
-         }
-
-         Log::info('analyze' . print_r($analyzeText, true));
-
-         if (isset($analyzeText['status']) && $analyzeText['status']) {
-            if (UserRole::query()->where($column, $user_id)->exists()) {
-               if ($analyzeText['type'] == 'links' or $analyzeText['type'] == 'words') {
-                  $violations = $this->updateUserViolations($user_id, $column);
-
-                  $userUpom = '';
-
-                  if ($violations->violations < 3) {
-                     $this->sendMessage($chat_id, "Пользователь {$userUpom}, вы нарушаете правила чата. Это {$violations->violations} предупреждение.", $column);
-                  } elseif ($violations->violations == 3) {
-                     $this->sendMessage($chat_id, "Пользователь {$userUpom}, это последнее предупреждение. Следующее нарушение приведёт к удалению.", $column);
-                  } elseif ($violations->violations > 3) {
-                     $this->sendMessage($chat_id, "Пользователь {$userUpom} был исключён за нарушение правила чата", $column);
-                     $this->kickUser($user_id, $column);
-                     UserViolation::query()->where('id', $violations->id)->delete();
-                  }
-
-                  $this->deleteMessage($message_id, $chat_id, $column);
-               }
-               if ($analyzeText['type'] == 'tag') {
-                  $this->deleteMessage($message_id, $chat_id, $column);
-               }
+         if ($text) {
+            if (!$sticker) {
+               $analyzeText = $this->analyzeText($text);
+            } else {
+               $analyzeText = [
+                  'status' => true,
+                  'type' => 'words'
+               ];
             }
-            if (isset($analyzeText['answer'])) {
-               $this->sendMessage($chat_id, $analyzeText['answer'], $message_id);
+
+            Log::info('analyze' . print_r($analyzeText, true));
+
+            if (isset($analyzeText['status']) && $analyzeText['status']) {
+               if (UserRole::query()->where($column, $user_id)->exists()) {
+                  if ($analyzeText['type'] == 'links' or $analyzeText['type'] == 'words') {
+                     $violations = $this->updateUserViolations($user_id, $column);
+
+                     $userUpom = $this->getUserInfo($user_id, $column);
+
+                     if ($violations->violations < 3) {
+                        $this->sendMessage($chat_id, "Пользователь {$userUpom}, вы нарушаете правила чата. Это {$violations->violations} предупреждение.", $column);
+                     } elseif ($violations->violations == 3) {
+                        $this->sendMessage($chat_id, "Пользователь {$userUpom}, это последнее предупреждение. Следующее нарушение приведёт к удалению.", $column);
+                     } elseif ($violations->violations > 3) {
+                        $this->sendMessage($chat_id, "Пользователь {$userUpom} был исключён за нарушение правила чата", $column);
+                        $this->kickUser($user_id, $column);
+                        UserViolation::query()->where('id', $violations->id)->delete();
+                     }
+
+                     $this->deleteMessage($message_id, $chat_id, $column);
+                  }
+                  if ($analyzeText['type'] == 'tag') {
+                     $this->deleteMessage($message_id, $chat_id, $column);
+                  }
+               }
+               if (isset($analyzeText['answer'])) {
+                  $this->sendMessage($chat_id, $analyzeText['answer'], $message_id);
+               }
             }
          }
       } else {
